third_party/WebKit/LayoutTests/fast/peerconnection/RTCPeerConnection-ice-expected.txt

Tests the RTCPeerConnection Ice functionality.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testExecutionOrderClosedConnection() did not throw exception.
PASS pc = new webkitRTCPeerConnection(null, null); did not throw exception.
PASS error.name is "InvalidStateError"
PASS error.toString() is "InvalidStateError: The RTCPeerConnection's signalingState is 'closed'."
PASS events is [1,2,3]
PASS iceConnectionState is completed
PASS pc.addIceCandidate(null, unexpectedCallback, unexpectedCallback).catch(expectedTypeError); did not throw exception.
PASS pc.addIceCandidate({candidate: "candidate"}, unexpectedCallback, unexpectedCallback).catch(expectedTypeError); did not throw exception.
PASS pc.addIceCandidate(iceCandidate, null, unexpectedCallback).catch(expectedTypeError); did not throw exception.
PASS pc.addIceCandidate(iceCandidate, unexpectedCallback, null).catch(expectedTypeError); did not throw exception.
PASS pc.addIceCandidate(iceCandidate, addIceCandidateSuccess1, unexpectedCallback); did not throw exception.
PASS error.name is "TypeError"
PASS expectedTypeError was called.
PASS error.name is "TypeError"
PASS expectedTypeError was called.
PASS error.name is "TypeError"
PASS expectedTypeError was called.
PASS error.name is "TypeError"
PASS expectedTypeError was called.
PASS addIceCandidateSuccess1 was called.
PASS pc.addIceCandidate({candidate: "candidate", sdpMid: 0}, addIceCandidateSuccess2, unexpectedCallback); did not throw exception.
PASS addIceCandidateSuccess2 was called.
PASS pc.addIceCandidate({candidate: "candidate", sdpMLineIndex: 0}, addIceCandidateSuccess3, unexpectedCallback); did not throw exception.
PASS addIceCandidateSuccess3 was called.
PASS pc.addIceCandidate({candidate: "candidate", sdpMid: 0, sdpMLineIndex: 0}, addIceCandidateSuccess4, unexpectedCallback); did not throw exception.
PASS addIceCandidateSuccess4 was called.
PASS iceConnectionState is closed.
PASS successfullyParsed is true

TEST COMPLETE
